+++ internal/commands/account/permissions/permissions.go
@@ -1,0 +1,14 @@
+package permissions
+
+import (
+	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands"
+)
+
+// BasePermissionsCommand creates the base 'permissions' command
+func BasePermissionsCommand() commands.Command {
+	return &permissionsCommand{commands.New("permissions", "Manage permissions")}
+}
+
+type permissionsCommand struct {
+	*commands.BaseCommand
+}